+++ tests/base/src/test/java/org/keycloak/tests/oid4vc/issuance/signing/OID4VCJwtAuthorizationDetailsFlowTest.java
@@ -27,8 +27,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
- * JWT-specific authorization details flow tests.
- * Extends the base class to inherit common test logic while providing JWT-specific implementations.
+ * JWT-specific authorization_details tests for pre-authorized_code grant.
  */
 @KeycloakIntegrationTest(config = OID4VCIssuerTestBase.VCTestServerConfig.class)
 public class OID4VCJwtAuthorizationDetailsFlowTest extends OID4VCAuthorizationDetailsFlowTestBase {
